Move TreeNodeRewritable for ExecutionPlan to the tree_node
